Reduce GUI compile time by building a GUI library (#7412)
* Call ensure_in_venv before all imports

* Create `gui` lib and link to `synergy` and test bins

* Move GUI IPC code to GUI lib

* Improve `AppConfig` members and change `minimizeOnClose` to `closeToTray`

* Group getters and setters

* Add missing `m_CloseToTray` load

* Fixed lint issue

* Add TODO comment for main window integ test

* Remove redundant comments and add override

* Remove dead code

* Update ChangeLog

* Fixed activation related compile errors

* Disable concurrency cancel jobs

* Disable fail-fast for matrix

+++ src/test/unittests/gui/QIpcClientTests.cpp
@@ -15,7 +15,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-#include "gui/src/QIpcClient.h"
+#include "gui/QIpcClient.h"
 
 #include <gmock/gmock.h>
 #include <gtest/gtest.h>
@@ -25,7 +25,7 @@ using testing::StrEq;
 
 class MockStream : public QDataStreamProxy {
 public:
-  MOCK_METHOD(int, writeRawData, (const char *, int), (override));
+  MOCK_METHOD(qint64, writeRawData, (const char *, qint64), (override));
 };
 
 TEST(QIpcClientTests, sendCommand_anyCommand_commandSent) {
